fix: use empty-dir volume to mount theme
The values provided in this example does not work with the latest
versions of the helm chart. This is because the bitnami keycloak chart
mounts its own volume at /opt/bitnami/keycloak/providers. The solution
is to mount the same volume into the init container and move the theme
into /app-providers-dir, which is then mounted into the main keycloak
container at /opt/bitnami/keycloak/providers. See the prepare-write-dirs
init container in the bitnami chart.

Signed-off-by: Matthew Fisher <matt.fisher@fermyon.com>